flytt vedtaksbrevToggelNavnSuffix til utilklasse
